Move ColorScheme to its own file / submodule
