+++ src/kind.rs
@@ -3,6 +3,7 @@
 use crate::common::*;
 
 mod camera_info;
+mod color_scheme;
 mod extension;
 mod format;
 mod frame_metadata;
@@ -11,28 +12,13 @@ mod stream_kind;
 mod timestamp_domain;
 
 pub use camera_info::Rs2CameraInfo;
+pub use color_scheme::ColorScheme;
 pub use extension::Rs2Extension;
 pub use format::Rs2Format;
 pub use frame_metadata::Rs2FrameMetadata;
 pub use option::Rs2Option;
 pub use stream_kind::Rs2StreamKind;
 pub use timestamp_domain::TimestampDomain;
-
-/// The enumeration of color schemes.
-#[repr(usize)]
-#[derive(FromPrimitive, Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum ColorScheme {
-    Jet = 0,
-    Classic = 1,
-    WhiteToBlack = 2,
-    BlackToWhite = 3,
-    Bio = 4,
-    Cold = 5,
-    Warm = 6,
-    Quantized = 7,
-    Pattern = 8,
-    Hue = 9,
-}
 
 /// The enumeration of persistence controls.
 #[repr(usize)]
